requirements.txt
adal==1.2.7
altair==5.1.2
annotated-types==0.6.0
antlr4-python3-runtime==4.13.2
anyio==3.7.1
applicationinsights==0.11.10
argcomplete==3.5.3
asttokens==2.4.1
attrs==23.1.0
azure-appconfiguration==1.7.1
azure-batch==15.0.0b1
azure-cli==2.70.0
azure-cli-core==2.70.0
azure-cli-telemetry==1.1.0
azure-common==1.1.28
azure-core==1.32.0
azure-cosmos==3.2.0
azure-data-tables==12.4.0
azure-datalake-store==1.0.0a0
azure-functions==1.21.3
azure-keyvault-administration==4.4.0b2
azure-keyvault-certificates==4.7.0
azure-keyvault-keys==4.9.0b3
azure-keyvault-secrets==4.7.0
azure-mgmt-advisor==9.0.0
azure-mgmt-apimanagement==4.0.0
azure-mgmt-appconfiguration==3.1.0
azure-mgmt-appcontainers==2.0.0
azure-mgmt-applicationinsights==1.0.0
azure-mgmt-authorization==4.0.0
azure-mgmt-batch==17.3.0
azure-mgmt-batchai==7.0.0b1
azure-mgmt-billing==6.0.0
azure-mgmt-botservice==2.0.0
azure-mgmt-cdn==12.0.0
azure-mgmt-cognitiveservices==13.5.0
azure-mgmt-compute==33.0.0
azure-mgmt-containerinstance==10.2.0b1
azure-mgmt-containerregistry==10.3.0
azure-mgmt-containerservice==34.1.0
azure-mgmt-core==1.5.0
azure-mgmt-cosmosdb==9.7.0
azure-mgmt-databoxedge==1.0.0
azure-mgmt-datalake-store==1.1.0b1
azure-mgmt-datamigration==10.0.0
azure-mgmt-dns==8.0.0
azure-mgmt-eventgrid==10.2.0b2
azure-mgmt-eventhub==10.1.0
azure-mgmt-extendedlocation==1.0.0b2
azure-mgmt-hdinsight==9.0.0b3
azure-mgmt-imagebuilder==1.3.0
azure-mgmt-iotcentral==10.0.0b2
azure-mgmt-iothub==3.0.0
azure-mgmt-iothubprovisioningservices==1.1.0
azure-mgmt-keyvault==10.3.0
azure-mgmt-loganalytics==13.0.0b4
azure-mgmt-managementgroups==1.0.0
azure-mgmt-maps==2.0.0
azure-mgmt-marketplaceordering==1.1.0
azure-mgmt-media==9.0.0
azure-mgmt-monitor==5.0.1
azure-mgmt-msi==7.0.0
azure-mgmt-mysqlflexibleservers==1.0.0b3
azure-mgmt-netapp==10.1.0
azure-mgmt-policyinsights==1.1.0b4
azure-mgmt-postgresqlflexibleservers==1.1.0b2
azure-mgmt-privatedns==1.0.0
azure-mgmt-rdbms==10.2.0b17
azure-mgmt-recoveryservices==3.0.0
azure-mgmt-recoveryservicesbackup==9.1.0
azure-mgmt-redhatopenshift==1.5.0
azure-mgmt-redis==14.5.0
azure-mgmt-resource==23.1.1
azure-mgmt-search==9.1.0
azure-mgmt-security==6.0.0
azure-mgmt-servicebus==8.2.1
azure-mgmt-servicefabric==2.1.0
azure-mgmt-servicefabricmanagedclusters==2.1.0b1
azure-mgmt-servicelinker==1.2.0b3
azure-mgmt-signalr==2.0.0b2
azure-mgmt-sql==4.0.0b20
azure-mgmt-sqlvirtualmachine==1.0.0b5
azure-mgmt-storage==22.1.0
azure-mgmt-synapse==2.1.0b5
azure-mgmt-trafficmanager==1.0.0
azure-mgmt-web==7.3.1
azure-monitor-query==1.2.0
azure-multiapi-storage==1.3.0
azure-storage-common==1.4.2
azure-synapse-accesscontrol==0.5.0
azure-synapse-artifacts==0.20.0
azure-synapse-managedprivateendpoints==0.4.0
azure-synapse-spark==0.7.0
bcrypt==4.3.0
blinker==1.7.0
branca==0.7.0
cached-property==1.5.2
cachetools==5.3.2
certifi==2023.7.22
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.0
contourpy==1.2.0
cryptography==44.0.2
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
Deprecated==1.2.18
distro==1.8.0
executing==2.0.1
fabric==3.2.2
fastapi==0.115.11
fitfile==1.1.8
fitparse==1.2.0
fonttools==4.45.0
garminconnect==0.2.25
garmindb==3.6.2
garth==0.5.2
gitdb==4.0.11
GitPython==3.1.40
greenlet==3.0.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
humanfriendly==10.0
idbutils==1.1.0
idna==3.4
importlib-metadata==6.8.0
iniconfig==2.1.0
invoke==2.2.0
ipykernel==6.26.0
ipyleaflet==0.17.4
ipython==8.17.2
ipywidgets==8.1.1
isodate==0.7.2
javaproperties==0.5.2
jedi==0.19.1
Jinja2==3.1.2
jmespath==1.0.1
jsondiff==2.0.0
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyterlab-widgets==3.0.9
kiwisolver==1.4.5
knack==0.11.0
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
microsoft-security-utilities-secret-masker==1.0.0b3
msal==1.31.2b1
msal-extensions==1.2.0
msrest==0.7.1
msrestazure==0.6.4.post1
nest-asyncio==1.5.8
numpy==1.26.2
oauthlib==3.2.2
objprint==0.2.3
openai==1.3.0
openai-token-counter==1.0.2
packaging==23.2
pandas==2.1.3
paramiko==3.5.1
parso==0.8.3
Pillow==10.1.0
pkginfo==1.12.1.2
platformdirs==4.0.0
pluggy==1.5.0
portalocker==2.10.1
prompt-toolkit==3.0.41
protobuf==4.25.1
psutil==5.9.6
pure-eval==0.2.2
py-deviceid==0.1.0
pyarrow==14.0.1
pycomposefile==0.0.32
pycparser==2.22
pydantic==2.5.1
pydantic_core==2.14.3
pydeck==0.8.1b0
PyGithub==1.59.1
Pygments==2.17.0
PyJWT==2.10.1
pymsalruntime==0.17.1
PyNaCl==1.5.0
pyodbc==5.2.0
pyOpenSSL==25.0.0
pyparsing==3.1.1
pyreadline3==3.5.4
PySocks==1.7.1
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.2
pyzmq==25.1.1
referencing==0.31.0
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rpds-py==0.13.1
scp==0.13.6
semver==2.13.0
setuptools==75.8.2
simplejson==3.20.1
six==1.16.0
smmap==5.0.1
snakemd==2.1.0
sniffio==1.3.0
SQLAlchemy==2.0.36
sshtunnel==0.1.5
stack-data==0.6.3
starlette==0.46.0
streamlit==1.28.2
tabulate==0.9.0
tcxfile==1.0.4
tenacity==8.2.3
tiktoken==0.5.2
toml==0.10.2
toolz==0.12.0
tornado==6.4.2
tqdm==4.66.5
traitlets==5.13.0
traittypes==0.2.1
typing_extensions==4.12.2
tzdata==2023.3
tzlocal==5.2
urllib3==2.1.0
uvicorn==0.34.0
validators==0.22.0
watchdog==3.0.0
wcwidth==0.2.10
websocket-client==1.3.3
widgetsnbextension==4.0.9
withings-sync==4.2.5
wrapt==1.17.2
xmltodict==0.14.2
xyzservices==2023.10.1
zipp==3.17.0
